CRAFTERCMS-1166 Changed /component/taxonomy to /taxonomy content type
