--- src/main/webapp/repo-bootstrap/global/blueprints/headless_blog/config/studio/content-types/component/taxonomy/config.xml
@@ -1,16 +1,0 @@
-<content-type name="/component/taxonomy" is-wcm-type="true">
-	<label>Taxonomy</label>
-	<form>/component/taxonomy</form>
-	<form-path>simple</form-path>
-	<model-instance-path>NOT-USED-BY-SIMPLE-FORM-ENGINE</model-instance-path>
-	<file-extension>xml</file-extension>
-	<content-as-folder>false</content-as-folder>
-	<previewable>false</previewable>
-	<noThumbnail>true</noThumbnail>
-	<image-thumbnail></image-thumbnail>
-	<paths>
-		<includes>
-			<pattern>/site/taxonomy/.*</pattern>
-		</includes>
-	</paths>
-</content-type>
