+++ src/main/webapp/repo-bootstrap/global/blueprints/website_editorial/site/taxonomy/categories.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <component>
-	<content-type>/component/taxonomy</content-type>	<display-template/>
+	<content-type>/taxonomy</content-type>	<display-template/>
 	<merge-strategy>inherit-levels</merge-strategy>
 	<objectGroupId>6497</objectGroupId>
 	<objectId>6497af67-bd6b-b1e2-5b3d-2d2559524fc0</objectId>
