+++ src/main/webapp/repo-bootstrap/global/blueprints/website_editorial/site/taxonomy/feature-icons.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <component>
-	<content-type>/component/taxonomy</content-type>	<display-template/>
+	<content-type>/taxonomy</content-type>	<display-template/>
 	<merge-strategy>inherit-levels</merge-strategy>
 	<objectGroupId>ff91</objectGroupId>
 	<objectId>ff915902-c3c7-d3d1-b828-17f9828f919f</objectId>
