+++ src/main/webapp/repo-bootstrap/global/blueprints/website_editorial/site/taxonomy/segments.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <component>
-	<content-type>/component/taxonomy</content-type>	<display-template/>
+	<content-type>/taxonomy</content-type>	<display-template/>
 	<merge-strategy>inherit-levels</merge-strategy>
 	<objectGroupId>8686</objectGroupId>
 	<objectId>868672ef-1b5e-7846-5c25-ac2929d2ff6e</objectId>
